switch to swc/jest
This switches the typescript transpiler used in jest from ts-jest which
uses tsc with swc. Due to some differences in how the resulting code
loads, the circular dependencies in the cluster store caused issues with
mocking, so the utilities on `ClustersOverview` were moved into their
own module.
