+++ webui/react/package.json
@@ -82,6 +82,7 @@
     ]
   },
   "devDependencies": {
+    "@swc/jest": "^0.2.24",
     "@testing-library/dom": "^9.0.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -140,7 +141,6 @@
     "stylelint-config-standard": "^22.0.0",
     "stylelint-order": "^4.1.0",
     "stylelint-scss": "^3.18.0",
-    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4",
     "vite": "^4.1.1",
