+++ webui/react/src/components/ResourcePoolCard.tsx
@@ -2,7 +2,7 @@ import React, { Suspense, useMemo } from 'react';
 
 import SlotAllocationBar from 'components/SlotAllocationBar';
 import { V1ResourcePoolTypeToLabel, V1SchedulerTypeToLabel } from 'constants/states';
-import { maxPoolSlotCapacity } from 'pages/Clusters/ClustersOverview';
+import { maxPoolSlotCapacity } from 'pages/Clusters/utils';
 import { paths } from 'routes/utils';
 import { V1ResourcePoolType, V1RPQueueStat, V1SchedulerType } from 'services/api-ts-sdk';
 import awsLogoOnDark from 'shared/assets/images/aws-logo-on-dark.svg';
